wrong datatype

+++ source/builtins/env_utils.c
@@ -6,7 +6,7 @@
 /*   By: jvan-hal <jvan-hal@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/05/09 15:57:34 by jvan-hal      #+#    #+#                 */
-/*   Updated: 2023/07/29 18:36:46 by jeroen        ########   odam.nl         */
+/*   Updated: 2023/07/29 18:43:26 by jeroen        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "ft_string.h"
 #include <stdlib.h>
 
-void	print_2d_arr(int fd, char **arr, int size)
+void	print_2d_arr(int fd, char **arr, size_t size)
 {
 	size_t	i;
